// frontend/src/api/models/getTradeSalesReturnsParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { GetTradeSalesReturnsOrderDir } from './getTradeSalesReturnsOrderDir'
import type { GetTradeSalesReturnsStatus } from './getTradeSalesReturnsStatus'

export type GetTradeSalesReturnsParams = {
  /**
   * Search term (return number, customer name, order number)
   */
  search?: string
  /**
   * Customer ID
   */
  customer_id?: string
  /**
   * Sales Order ID
   */
  sales_order_id?: string
  /**
   * Warehouse ID
   */
  warehouse_id?: string
  /**
   * Return status
   */
  status?: GetTradeSalesReturnsStatus
  /**
   * Multiple return statuses
   */
  statuses?: string[]
  /**
   * Start date (ISO 8601)
   */
  start_date?: string
  /**
   * End date (ISO 8601)
   */
  end_date?: string
  /**
   * Minimum refund amount
   */
  min_amount?: number
  /**
   * Maximum refund amount
   */
  max_amount?: number
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
  /**
   * Order by field
   */
  order_by?: string
  /**
   * Order direction
   */
  order_dir?: GetTradeSalesReturnsOrderDir
}