// frontend/src/api/models/getTradeSalesReturnsStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type GetTradeSalesReturnsStatus =
  (typeof GetTradeSalesReturnsStatus)[keyof typeof GetTradeSalesReturnsStatus]

export const GetTradeSalesReturnsStatus = {
  DRAFT: 'DRAFT',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const